+++ tests/tasks_v2/test_illumination_correction_task.py
@@ -462,7 +462,7 @@ def test_multidimensional_input(
     )
 
     table = origin_ome_zarr.build_image_roi_table("image")
-    origin_ome_zarr.add_table("well_ROI_table", table, backend="experimental_json_v1")
+    origin_ome_zarr.add_table("well_ROI_table", table, backend="anndata")
 
     # Prepare arguments for illumination_correction function
     testdata_str = testdata_path.as_posix()